Moved source files to src, babel to lib, entry to lib
