+++ .gitignore
@@ -1,4 +1,5 @@
 .idea/
+lib/
 
 # Logs
 logs
