+++ package.json
@@ -2,14 +2,16 @@
   "name": "yttrium-server",
   "version": "0.1.2",
   "description": "The jQuery web server framework",
-  "main": "index.js",
+  "main": "lib/index.js",
   "directories": {
     "test": "test"
   },
   "scripts": {
     "lint": "eslint .",
     "test": "mocha",
-    "coveralls": "nyc -a npm test && nyc report --reporter=text-lcov | coveralls"
+    "coveralls": "nyc -a npm test && nyc report --reporter=text-lcov | coveralls",
+    "compile": "babel --presets es2015 -d lib/ src/",
+    "prepublish": "npm test && npm run compile"
   },
   "repository": {
     "type": "git",
@@ -30,6 +32,8 @@
     "jsdom": "^11.1.0"
   },
   "devDependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
     "chai": "^4.1.1",
     "chai-as-promised": "^7.1.1",
     "coveralls": "^2.13.1",
