+++ test/router.spec.js
@@ -2,7 +2,7 @@
 const chai = require('chai');
 const sinon = require('sinon');
 const sinonChai = require('sinon-chai');
-const Router = require('../router');
+const Router = require('../src/router');
 
 const expect = chai.expect;
 chai.use(sinonChai);
